complete changing App layout
