+++ app/imports/ui/pages/ListStuff.jsx
@@ -35,10 +35,6 @@ class ListStuff extends React.Component {
     this.setState({ gender: e.target.value });
   };
 
-  componentDidMount() {
-    // console.log(this.props);
-  }
-
   submit = () => {
     const { _id } = this.props.info;
     const { gender, vaccinated, vaccineType, vaccineLot } = this.state;
@@ -54,7 +50,6 @@ class ListStuff extends React.Component {
 
   componentDidUpdate(prevProps) {
     if (this.props.info && this.props.info._id !== prevProps.info?._id) {
-      console.log("hi");
       this.setState({
         name: this.props.info.name,
         gender: this.props.info.gender,
@@ -67,8 +62,8 @@ class ListStuff extends React.Component {
   // Render the page once subscriptions have been received.
   renderPage() {
     return (
-      <Container id="signup-page">
-        <Grid textAlign="center" verticalAlign="middle" centered columns={2}>
+      <div id="signup-page" style={{width: "30rem"}}>
+        <Grid textAlign="center" verticalAlign="middle" centered>
           <Grid.Column>
             <Header as="h2" textAlign="center">
               Edit my info
@@ -137,7 +132,7 @@ class ListStuff extends React.Component {
             </Form>
           </Grid.Column>
         </Grid>
-      </Container>
+      </div>
     );
   }
 }
